+++ src/components/Select/selectChangelogData.ts
@@ -10,6 +10,13 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "2026-03-05",
+    version: "4.1.5",
+    type: "Update",
+    affects: ["Styles"],
+    notes: ["Fixes the RTL styles for the dropdown arrow."],
+  },
+  {
     date: "2025-08-11",
     version: "4.0.0",
     type: "Update",